unconditionally enable ROOT MT safety
Now that the bulk of the work is done by spawned threads, we must always
enable safe MT, even if we only have a single worker.

This would not be required if all work was done on the main thread when
using a single worker, but this is not the case.