Added setting for `max_charge_bunches` in [TransientPropagation], to reduce the negative effects of unexpectedly large energy deposits.
The setting allows a user to decide on a maximum number of charge bunches that can be propagated form a single deposit. Setting this overrides and increases `charge_per_step` for any deposit that exceeds the product of `charge_per_step` and `max_charge_bunches`. This speeds up handling of large deposits, while reducing the simulation precision for them.