Fixed (hacked) it so that rotations of scalar fields work even if the x coordinate is taken to be TCAD y or z.
This is a quick fix, will see if I can make it more elegant. It currently duplicates the field value to always have it in the x-position. The issue appears when the final scalar value ends up in a position different from x in the final Point object, as only the x-value of the point is used to save scalar fields after parsing.