Compatibility with CsPbBr3 (Perovskites) as active material
This contribution adds Cesium Lead Bromide (CsPbBr3) to the list of
active materials. This material is being extensivelly investigated as
photo-active sensor for hard X-ray detection, so we made a few changes
to make Allpix Squared compatible to this material. While the Fano
factor has not yet been reported (assumed to equal 0.1), other
properties (such as ionization energy of 5.3 eV) are implemented here.
Carrier mobility can be simulated using the consant_mobility model, with
 values ranging from 11 to 128 cm2/Vs for holes and from 28 to 160
cm2/Vs for electrons.

https://doi.org/10.1002/adom.201600704
https://www.nature.com/articles/s41467-021-21805-0
https://doi.org/10.1038/s41598-024-74384-7